pyproject.toml
[project]
name = "win-voice-injector"
version = "0.1.0"
description = "Windows 常駐型音声入力アプリ"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "google-genai>=1.64.0",
    "keyboard>=0.13.5",
    "numpy>=2.4.2",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]

[project.scripts]
voice-injector = "main:main"
